+++ test/integration/tests/5180-ghc-rts-flags/files/src/Lib.hs
@@ -1,0 +1,4 @@
+module Lib where
+
+main :: IO ()
+main = putStrLn "hello world"